refactor: rename action.js to call.js
