+++ lib/tools.js
@@ -20,7 +20,7 @@ const {
   registerPerActionTools,
   executeCallActionTool,
   executePerActionTool
-} = require('./tools/action')
+} = require('./tools/call')
 
 module.exports = {
   // Definition factories (shared with compile.js)
